+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.2.2
 
 * Support incremental annotation processing, see #48
+* Omit the type for private inner classes when using reflection, see #44
 
 ## 1.2.1 (2018-03-18)